chore(backup): move runner under scripts/backup with README
Colocate the snapshot runner with design/usage docs; update systemd,
ops runbook, and test imports to the new path.
